+++ core/src/tunnel.cpp
@@ -921,7 +921,7 @@ static std::shared_ptr<ServerUpstream> select_upstream(
         }
         if (const SocketAddress *dst; // NOLINT(cppcoreguidelines-init-variables)
                 conn != nullptr && conn->flags.test(CONNF_LOOKINGUP_DOMAIN)
-                && (conn->flags.test(CONNF_SUSPECT_EXCLUSION) || self->vpn->exclusions_tcp_early_ack)
+                && (conn->flags.test(CONNF_SUSPECT_EXCLUSION) || self->vpn->exclusions_tcp_early_ack_enabled)
                 && nullptr != (dst = std::get_if<SocketAddress>(&conn->addr.dst))
                 && is_domain_scannable_port(dst->port())) {
             return self->fake_upstream;