+++ trusttunnel/include/vpn/trusttunnel/config.h
@@ -57,7 +57,7 @@ struct TrustTunnelConfig {
     bool killswitch_enabled = false;
     std::string killswitch_allow_ports;
     bool post_quantum_group_enabled = true;
-    bool exclusions_tcp_early_ack = false;
+    bool exclusions_tcp_early_ack_enabled = false;
     bool exclusions_preresolve_enabled = true;
     uint32_t exclusions_preresolve_max_queries = 50;
     std::string log_file_path;